WaterGuanyin/ai_business_card/requirements.txt
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
fake-headers==1.0.2
html5lib==1.0.1
idna==2.8
lxml==4.4.1
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
urllib3==1.25.7
webencodings==0.5.1
pyquery==1.4.1
feedparser==5.2.1
flask==1.1.1
flask_cors==3.0.8
line-bot-sdk==1.15.0
Pillow==6.2.1